ItemStackDTO is now an interface, and Deserializer uses ConfigUtils.GSON
